+++ control/magma_threadsetting.cpp
@@ -80,8 +80,8 @@ magma_int_t magma_get_parallel_numthreads()
         threads = strtol( threads_str, &endptr, 10 );
         if ( threads < 1 || *endptr != '\0' ) {
             threads = 1;
-            fprintf( stderr, "$MAGMA_NUM_THREADS='%s' is an invalid number; using %ld threads.\n",
-                     threads_str, long(threads) );
+            fprintf( stderr, "$MAGMA_NUM_THREADS='%s' is an invalid number; using %lld threads.\n",
+                     threads_str, (long long) threads );
         }
     }
     else {
